Initial commit
A frequency-sorted file of passwords (specified by FILENAME in
store_passwords.py) is divided into segments and stored in an array of
bloom filters. Bloom filter size is linear in the number of elements, so
segmentation doesn't waste significant space.

Once these filters are built, evaluate.py can test a password's strength
by looking for the first bloom filter that contains it. The number of
passwords that were in previous bloom filters gives an estimate of how
many hashes a realistic attacker -- as realistic as the sorted passwords
file -- would need to calculate before cracking it.

Bloom filter submodule will be added in next commit.