Logarithmically scaled filter sizes
There are now many small filters and a few big ones since the differnce between 'one second to crack' and 'one hour to crack' is much bigger than the difference between 'one hour to crack' and 'two hours to crack'.

Filters are now stored in an ordered dict with each key equivilent to the total number of hashes in the previous filters (including the one matching the key). The largest key is the total password count.

Filters are a bit more efficient now: the full capacity -- based on the allocated space, not the target size -- of each filter is used to fill it.

Cracking-times are now lower-bounded at 1 sec to avoid saying 'less than 0 seconds'.